extract bp tests into separate file
